requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
chromadb>=0.4.22
beautifulsoup4>=4.12.3
requests>=2.31.0
python-dotenv>=1.0.1
numpy>=1.26.0
langchain-text-splitters>=0.0.2
lxml>=5.1.0
tiktoken>=0.6.0
